# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ProductHub API — Enterprise Product, Inventory & Order Management Platform",
    "version": "1.0.0",
    "description": "Production-grade enterprise platform engineered with Go (Fiber), MySQL (GORM), gRPC internal service mesh, Redis Cache-Aside with Sliding Window Rate Limiting and Idempotent Mutations, and Asynchronous Worker Pools.",
    "contact": {
      "name": "ProductHub Engineering Team",
      "url": "https://github.com/producthub/platform"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Local HTTP API Gateway"
    }
  ],
  "tags": [
    { "name": "Health", "description": "System health and dependency diagnostics" },
    { "name": "Authentication", "description": "User registration, JWT authentication & profile" },
    { "name": "Categories", "description": "Product catalog classification" },
    { "name": "Products", "description": "Catalog management with Redis Cache-Aside" },
    { "name": "Inventory", "description": "Stock levels, pessimistic concurrency reservation & restocking" },
    { "name": "Cart", "description": "Customer shopping cart operations" },
    { "name": "Orders", "description": "ACID transactional checkout with Idempotency & state transitions" },
    { "name": "Admin Operations", "description": "Administrative management & order workflows" },
    { "name": "Worker Pool", "description": "Background goroutine worker pool telemetry" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token obtained from /api/auth/login"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" },
          "meta": { "type": "object" }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "BAD_REQUEST" },
              "message": { "type": "string", "example": "Invalid payload provided" }
            }
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "example": "John Doe" },
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "minLength": 6, "example": "Secure@123" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "customer@example.com" },
          "password": { "type": "string", "example": "Customer@123" }
        }
      },
      "ProductRequest": {
        "type": "object",
        "required": ["title", "sku", "category_id", "price", "stock"],
        "properties": {
          "title": { "type": "string", "example": "Mechanical Gaming Keyboard RGB" },
          "sku": { "type": "string", "example": "KB-RGB-001" },
          "category_id": { "type": "integer", "example": 1 },
          "description": { "type": "string", "example": "Hot-swappable tactile switches with per-key RGB backlighting" },
          "price": { "type": "number", "format": "float", "example": 89.99 },
          "stock": { "type": "integer", "example": 50 },
          "low_stock_threshold": { "type": "integer", "example": 10 }
        }
      },
      "RestockRequest": {
        "type": "object",
        "required": ["product_id", "quantity"],
        "properties": {
          "product_id": { "type": "integer", "example": 1 },
          "quantity": { "type": "integer", "minimum": 1, "example": 50 },
          "reason": { "type": "string", "example": "Warehouse replenishment PO-8834" }
        }
      },
      "CartItemRequest": {
        "type": "object",
        "required": ["product_id", "quantity"],
        "properties": {
          "product_id": { "type": "integer", "example": 1 },
          "quantity": { "type": "integer", "minimum": 1, "example": 2 }
        }
      },
      "CheckoutRequest": {
        "type": "object",
        "required": ["shipping_address", "payment_method"],
        "properties": {
          "shipping_address": { "type": "string", "example": "123 Innovation Way, Suite 400, Tech District" },
          "payment_method": { "type": "string", "enum": ["CREDIT_CARD", "DEBIT_CARD", "UPI", "COD"], "example": "CREDIT_CARD" }
        }
      },
      "OrderStatusUpdateRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "enum": ["CONFIRMED", "PROCESSING", "SHIPPED", "DELIVERED", "CANCELLED"],
            "example": "CONFIRMED"
          }
        }
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": ["Health"],
        "summary": "System Health & Dependency Check",
        "description": "Performs live ping against MySQL database and returns operational status.",
        "responses": {
          "200": { "description": "All systems operational" },
          "503": { "description": "Database disconnected or degraded" }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register New Customer Account",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } }
        },
        "responses": {
          "201": { "description": "User created with JWT token" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Authentication & JWT Issuance",
        "description": "Rate limited to 15 attempts/min per IP to prevent brute-force attacks.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
        },
        "responses": {
          "200": { "description": "Login successful, returns JWT bearer token" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get Current Authenticated User Profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "List All Product Categories",
        "responses": { "200": { "description": "List of categories" } }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create Category (Admin Only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Audio & Electronics" },
                  "description": { "type": "string", "example": "Headphones, speakers, and peripherals" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created" },
          "403": { "description": "Forbidden — ADMIN role required" }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "List Products with Pagination, Search & Filters",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "category_id", "in": "query", "schema": { "type": "integer" } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "default": "created_at" } },
          { "name": "sort_order", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" } }
        ],
        "responses": { "200": { "description": "Paginated list of products" } }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create New Product (Admin Only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } } }
        },
        "responses": {
          "201": { "description": "Product created" },
          "403": { "description": "Forbidden — ADMIN role required" }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get Product Details (Redis Cache-Aside)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "Product record. Includes X-Cache: HIT/MISS response header.",
            "headers": { "X-Cache": { "schema": { "type": "string", "example": "HIT" } } }
          },
          "404": { "description": "Product not found" }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update Product (Admin Only)",
        "description": "Automatically invalidates product Redis cache keys on mutation.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProductRequest" } } }
        },
        "responses": {
          "200": { "description": "Product updated" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Deactivate Product (Admin Only)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Product deactivated" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/inventory/summary": {
      "get": {
        "tags": ["Inventory"],
        "summary": "Inventory KPIs & Aggregate Stock Summary (Admin Only)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Total stock, out of stock, low stock counts" } }
      }
    },
    "/api/inventory/restock": {
      "post": {
        "tags": ["Inventory"],
        "summary": "Restock Product Inventory Units (Admin Only)",
        "description": "Uses pessimistic locking (SELECT FOR UPDATE) to safely replenish warehouse stock and invalidates Redis cache.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RestockRequest" } } }
        },
        "responses": {
          "200": { "description": "Stock replenished" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get Current Customer's Shopping Cart",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Cart items with computed subtotal and tax" } }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Clear All Items from Cart",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "Cart cleared" } }
      }
    },
    "/api/cart/items": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add Item to Shopping Cart",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CartItemRequest" } } }
        },
        "responses": { "200": { "description": "Item added to cart" } }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List Customer's Placed Orders",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "List of user orders" } }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Checkout Cart into Placed Order (ACID & Idempotent)",
        "description": "Atomically locks stock via SELECT FOR UPDATE, deducts inventory, empties cart, creates order, and dispatches ORDER_CREATED event to worker pool. Protected with client Idempotency-Key.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string", "format": "uuid" },
            "description": "RFC-4122 v4 UUID preventing duplicate order creation on network retries."
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CheckoutRequest" } } }
        },
        "responses": {
          "201": { "description": "Order created successfully" },
          "200": { "description": "Idempotent replay of previously completed checkout" },
          "409": { "description": "Concurrent checkout mutation currently processing for this key" },
          "422": { "description": "Insufficient inventory stock for an item" }
        }
      }
    },
    "/api/orders/{id}/cancel": {
      "post": {
        "tags": ["Orders"],
        "summary": "Cancel Order and Restock Inventory (Customer / Admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Order cancelled and reserved stock restored" },
          "422": { "description": "Order cannot be cancelled in its current state" }
        }
      }
    },
    "/api/admin/orders": {
      "get": {
        "tags": ["Admin Operations"],
        "summary": "List All Orders Across the Platform (Admin Only)",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "List of all orders" } }
      }
    },
    "/api/admin/orders/{id}/status": {
      "put": {
        "tags": ["Admin Operations"],
        "summary": "Transition Order State in State Machine (Admin Only)",
        "description": "Enforces valid transition paths: PENDING -> CONFIRMED -> PROCESSING -> SHIPPED -> DELIVERED, or CANCELLED.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderStatusUpdateRequest" } } }
        },
        "responses": {
          "200": { "description": "Status updated and ORDER_STATUS_CHANGED event dispatched" },
          "422": { "description": "Invalid status transition" }
        }
      }
    },
    "/api/admin/workers/stats": {
      "get": {
        "tags": ["Worker Pool"],
        "summary": "Live Diagnostics for Background Worker Pool (Admin Only)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Queue depth, worker count, total processed & failed counts" }
        }
      }
    }
  }
}